requirements.txt
fastapi
uvicorn
python-multipart
python-decouple
langchain
langchain-community
langchain-text-splitters
chromadb
pypdf
groq
scikit-learn
numpy